requirements.txt
statsmodels==0.11.1
xlrd==1.2.0
typer==0.3.1
sqlalchemy==1.3.19
pandas==1.1.1
seaborn==0.10.1
scipy==1.5.2
reportlab==3.5.26
pyodbc==4.0.30
parse==1.19.0
openpyxl==3.0.5
numpy==1.19.1
matplotlib==3.3.1
cx_oracle==7.2.3
filterpy==1.4.5
python-dotenv==0.14.0
PyCygNet==0.0.6
toml==0.10.1